refactor(bookmarks): split bookmark shell and API like other routes
Defer browser-only ProfileService UI behind app-bookmark-api so SSR never constructs timers or auth-coupled bookmark logic.
