+++ cultpodcasts/src/app/bookmark-api/bookmark-api.component.html
@@ -1,0 +1,11 @@
+@if (isAuthenticated()) {
+<button mat-icon-button (click)="bookmark()">
+    @if (waitingCallback() && !bookmarkTimeout()) {
+    <mat-spinner></mat-spinner>
+    } @else {
+    <mat-icon [class.material-symbols-outlined]="!isBookmarked()">
+        bookmark
+    </mat-icon>
+    }
+</button>
+}
