+++ cultpodcasts/src/app/bookmark-api/bookmark-api.component.sass
@@ -1,0 +1,18 @@
+.mdc-icon-button
+    position: absolute
+    top: 12px
+    right: 12px
+    margin-right: 0
+    mat-spinner
+        width: 24px !important
+        height: 24px  !important
+@media screen and (max-width: 400px)
+    .mdc-icon-button
+        right: 0
+:host.has-menu
+    .mdc-icon-button
+        margin-right: 44px
+@media screen and (max-width: 400px)
+    :host.has-menu
+        .mdc-icon-button
+            margin-right: 36px
