+++ cultpodcasts/src/app/bookmark-api/bookmark-api.component.ts
@@ -1,0 +1,83 @@
+import {
+  ChangeDetectionStrategy,
+  Component,
+  DestroyRef,
+  HostBinding,
+  OnInit,
+  computed,
+  inject,
+  input,
+  signal
+} from '@angular/core';
+import { takeUntilDestroyed, toSignal } from '@angular/core/rxjs-interop';
+import { ProfileService } from '../profile.service';
+import { MatIconModule } from '@angular/material/icon';
+import { MatButtonModule } from '@angular/material/button';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+
+@Component({
+  selector: 'app-bookmark-api',
+  imports: [
+    MatIconModule,
+    MatButtonModule,
+    MatProgressSpinnerModule
+  ],
+  templateUrl: './bookmark-api.component.html',
+  styleUrl: './bookmark-api.component.sass',
+  changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class BookmarkApiComponent implements OnInit {
+  episodeId = input.required<string>();
+  hasMenu = input<boolean>(false);
+  protected readonly waitingCallback = signal(true);
+  protected readonly bookmarkTimeout = signal(false);
+  protected readonly isAuthenticated = toSignal(
+    inject(ProfileService).isAuthenticated$,
+    { initialValue: false }
+  );
+  private readonly profileService = inject(ProfileService);
+  private readonly destroyRef = inject(DestroyRef);
+  private readonly bookmarks = toSignal(
+    this.profileService.bookmarks$,
+    { initialValue: new Set<string>() }
+  );
+  private readonly isBookmarkedState = signal<boolean | undefined>(undefined);
+  protected readonly isBookmarked = computed(() =>
+    this.bookmarks().has(this.episodeId())
+  );
+
+  @HostBinding('class.has-menu')
+  get hasMenuGet() { return this.hasMenu(); }
+
+  constructor() {
+    setTimeout(() => this.bookmarkTimeout.set(true), 5000);
+  }
+
+  ngOnInit(): void {
+    // Required input is only safe after construction (NG0950 if read in constructor
+    // when bookmarks$ ReplaySubject already has a value and emits synchronously).
+    this.profileService.bookmarks$
+      .pipe(takeUntilDestroyed(this.destroyRef))
+      .subscribe(bookmarks => {
+        const state = bookmarks.has(this.episodeId());
+        if (this.waitingCallback() && state !== this.isBookmarkedState()) {
+          this.waitingCallback.set(false);
+        }
+        this.isBookmarkedState.set(state);
+      });
+  }
+
+  async bookmark(): Promise<void> {
+    const bookmarked = this.profileService.bookmarks.has(this.episodeId());
+    this.waitingCallback.set(true);
+    try {
+      if (bookmarked) {
+        await this.profileService.removeBookmark(this.episodeId());
+      } else {
+        await this.profileService.addBookmark(this.episodeId());
+      }
+    } finally {
+      this.waitingCallback.set(false);
+    }
+  }
+}
