+++ cultpodcasts/src/app/bookmark/bookmark.component.ts
@@ -1,93 +1,14 @@
-import {
-  ChangeDetectionStrategy,
-  Component,
-  DestroyRef,
-  HostBinding,
-  OnInit,
-  PLATFORM_ID,
-  Signal,
-  computed,
-  inject,
-  input,
-  signal
-} from '@angular/core';
-import { takeUntilDestroyed, toSignal } from '@angular/core/rxjs-interop';
-import { isPlatformBrowser } from '@angular/common';
-import { ProfileService } from '../profile.service';
-import { MatIconModule } from '@angular/material/icon';
-import { MatButtonModule } from '@angular/material/button';
-import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+import { ChangeDetectionStrategy, Component, input } from '@angular/core';
+import { BookmarkApiComponent } from '../bookmark-api/bookmark-api.component';
 
 @Component({
   selector: 'app-bookmark',
-  imports: [
-    MatIconModule,
-    MatButtonModule,
-    MatProgressSpinnerModule
-  ],
+  imports: [BookmarkApiComponent],
   templateUrl: './bookmark.component.html',
   styleUrl: './bookmark.component.sass',
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class BookmarkComponent implements OnInit {
+export class BookmarkComponent {
   episodeId = input.required<string>();
   hasMenu = input<boolean>(false);
-  protected readonly waitingCallback = signal(true);
-  /** Browser-only: Zone-tracked timers deadlock SSR on Cloudflare Workers. */
-  protected readonly bookmarkTimeout: Signal<boolean> = timerSignal(5000);
-  protected readonly isAuthenticated = toSignal(
-    inject(ProfileService).isAuthenticated$,
-    { initialValue: false }
-  );
-  private readonly profileService = inject(ProfileService);
-  private readonly destroyRef = inject(DestroyRef);
-  private readonly bookmarks = toSignal(
-    this.profileService.bookmarks$,
-    { initialValue: new Set<string>() }
-  );
-  private readonly isBookmarkedState = signal<boolean | undefined>(undefined);
-  protected readonly isBookmarked = computed(() =>
-    this.bookmarks().has(this.episodeId())
-  );
-
-  @HostBinding('class.has-menu')
-  get hasMenuGet() { return this.hasMenu(); }
-
-  ngOnInit(): void {
-    // Required input is only safe after construction (NG0950 if read in constructor
-    // when bookmarks$ ReplaySubject already has a value and emits synchronously).
-    this.profileService.bookmarks$
-      .pipe(takeUntilDestroyed(this.destroyRef))
-      .subscribe(bookmarks => {
-        const state = bookmarks.has(this.episodeId());
-        if (this.waitingCallback() && state !== this.isBookmarkedState()) {
-          this.waitingCallback.set(false);
-        }
-        this.isBookmarkedState.set(state);
-      });
-  }
-
-  async bookmark(): Promise<void> {
-    const bookmarked = this.profileService.bookmarks.has(this.episodeId());
-    this.waitingCallback.set(true);
-    try {
-      if (bookmarked) {
-        await this.profileService.removeBookmark(this.episodeId());
-      } else {
-        await this.profileService.addBookmark(this.episodeId());
-      }
-    } finally {
-      this.waitingCallback.set(false);
-    }
-  }
-}
-
-function timerSignal(ms: number): Signal<boolean> {
-  const done = signal(false);
-  if (isPlatformBrowser(inject(PLATFORM_ID))) {
-    setTimeout(() => done.set(true), ms);
-  } else {
-    done.set(true);
-  }
-  return done.asReadonly();
 }
